chore: move `mul_with_witness` into `generated_acir.rs`
